First pass at cleaning up Application vs Runner.
The Runner class is now responsible for the options parsing. This delegates
the Application to _only_ loading up an application based on file name. This
simplifes what's needed in order to create a custom server.

Added some documentation to the various classes to make stuff more clear for
other developers.