root_path doesn't always do the right thing.
This fix is a bit of a hack. If we haven't exeuted the runner (ie, we're parsing the file)
then we haven't chdir'd to the API directory, which means we need a different meaning for
root_path. This fix adds an app_path method which can be used prior to the chdir. Root_path
can be used after the chdir. Like I said, hack. Can't think of a better way to handle
this at the moment without storing a global specifying if we've changed directories or not.

Fixes #40.