Work around Rack::Static bug.
If the :index isn't provided it defaults to index.html. There is no way, in
the current gem version, to set :index to not trigger, so you can't have
a route on / when using Rack::Static. This is fixed in Rack master, just
need a new rack version pushed and we can change back to '/'.