Move all environment handling into the runner.
Previously, the environment handling was performed in a couple of
different places because of a need for it to be set at require
time, but be parsed during runtime.  This commit doesn't change
that behaviour, but does tidy it all up by encapsulating the
environment parsing within the runner.

This fixes the issue whereby the environment was always reset by
the command-line parsing in the runner, even if it had already
been set via RACK_ENV.

It also enforces a precedence order for where the environment
variable can come from.  In increasing order of precedence:

1. Default (as set above)
2. RACK_ENV
3. -e/--environment command line options
4. Hard-coded call to Goliath#env=