Modify DefaultResponseFormat middleware to support dynamic response bodies
We mustn't assume that the user wants to load the entire response body
into memory or at once, and instead allow them return a response body
object that responds to #each and dynamically generates chunks of the
response body.

In these cases the AsyncRack::ContentLength middleware will not be able
to calculate `Content-Length`, but AsyncRack::ContentLength is just a
wrapper around Rack::ContentLength, which has the exact same behaviour,
so Rack users should expect that.