Improve handling multiple requests to same connection
It is possible that a new request is being received on the same TCP
socket before the response was even sent for the previous request. In
this case we must make sure that we're not interleaving two responses;
the response for the second request can only be sent once the first
response was sent.

The logic for handling this wasn't quite right. When request is
terminated, Goliath would check if there are any pending requests,
promote the first one to "current" and mark it as ready. However, that
means that request that was "current" at that moment is never closed.

So we update the logic to first finalize any "current" requests by
closing them, and then promoting any pending requests to "current". We
also stop requests from marking themselves as succeeded on exceptions,
because in that case it's theoretically possible that a second request
starts sending its response before the first request even started.