Stream the response to the client
EventMachine will buffer all data that was sent to the connection in the
same tick, and actually write it to the TCP socket only at the end of
the tick. That means that if the response body is large and/or it takes
time to generate, Goliath will wait until it has been loaded whole
before sending it to the client.

This is inefficient. First, it leads to high memory consumption on large
response bodies, because the content is buffered in-memory before it is
sent to the client. Second, if the response body takes time to generate,
the client needs to wait until the end before it can start downloading
the response body.

Both these problems are fixed by sending chunks of the response body as
it is being generated. This is done by sending each chunk in a new EM
tick, which prevents EM from buffering the response.