Stop sending response data if connection is closed
If the connection has been closed, there is no point in continuing
iterating over the response body and continuing to send data, because it
will never get written to the socket anyway. This saves on performance.

Also, once `#close` has been called on the response body, continuing to
iterate over it might raise an exception. Imagine a scenario where the
body is a lazy enumerable that retrieves chunks from an open file, and
`#close` closes the file.